pindahkan table to pages/
